chore: Organize components
