+++ app/resources/js/components/laravel/app-content.tsx
