+++ app/resources/js/components/laravel/app-logo-icon.tsx
