+++ app/resources/js/components/laravel/app-logo.tsx
