+++ app/resources/js/components/laravel/app-shell.tsx
