+++ app/resources/js/components/laravel/heading-small.tsx
