+++ app/resources/js/components/laravel/heading.tsx
