+++ app/resources/js/components/laravel/icon.tsx
