+++ app/resources/js/components/laravel/input-error.tsx
