+++ app/resources/js/components/laravel/logo.tsx
