+++ app/resources/js/components/laravel/nav-main.tsx
