+++ app/resources/js/components/laravel/system-management-card.tsx
