+++ app/resources/js/components/laravel/user-info.tsx
