+++ app/resources/js/layouts/app/app-header-layout.tsx
@@ -1,18 +1,18 @@
-import { AppContent } from '@/components/app-content';
-import { AppHeader } from '@/components/app-header';
 import { AppShell } from '@/components/app-shell';
+import { AppContent } from '@/components/laravel/app-content';
+import { AppHeader } from @/components/laravel/app-shellapp-header';
 import { type BreadcrumbItem } from '@/types';
 
 interface AppHeaderLayoutProps {
-    children: React.ReactNode;
-    breadcrumbs?: BreadcrumbItem[];
+  children: React.ReactNode;
+  breadcrumbs?: BreadcrumbItem[];
 }
 
 export default function AppHeaderLayout({ children, breadcrumbs }: AppHeaderLayoutProps) {
-    return (
-        <AppShell>
-            <AppHeader breadcrumbs={breadcrumbs} />
-            <AppContent>{children}</AppContent>
-        </AppShell>
-    );
+  return (
+    <AppShell>
+      <AppHeader breadcrumbs={breadcrumbs} />
+      <AppContent>{children}</AppContent>
+    </AppShell>
+  );
 }
