+++ app/resources/js/pages/admin/gestao-sistema/Tt-icon.tsx
@@ -1,0 +1,7 @@
+export function TtIcon() {
+  return (
+    <svg width="20" height="16" viewBox="0 0 20 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M5 16V3H0V0H13V3H8V16H5ZM14 16V8H11V5H20V8H17V16H14Z" fill="#094785" />
+    </svg>
+  );
+}
