+++ app/resources/js/pages/auth/verify-email.tsx
@@ -3,39 +3,42 @@ import { Head, useForm } from '@inertiajs/react';
 import { LoaderCircle } from 'lucide-react';
 import { FormEventHandler } from 'react';
 
-import TextLink from '@/components/text-link';
+import TextLink from '@/components/laravel/text-link';
 import { Button } from '@/components/ui/button';
 import AuthLayout from '@/layouts/auth-layout';
 
 export default function VerifyEmail({ status }: { status?: string }) {
-    const { post, processing } = useForm({});
+  const { post, processing } = useForm({});
 
-    const submit: FormEventHandler = (e) => {
-        e.preventDefault();
+  const submit: FormEventHandler = (e) => {
+    e.preventDefault();
 
-        post(route('verification.send'));
-    };
+    post(route('verification.send'));
+  };
 
-    return (
-        <AuthLayout title="Verify email" description="Please verify your email address by clicking on the link we just emailed to you.">
-            <Head title="Email verification" />
+  return (
+    <AuthLayout
+      title="Verify email"
+      description="Please verify your email address by clicking on the link we just emailed to you."
+    >
+      <Head title="Email verification" />
 
-            {status === 'verification-link-sent' && (
-                <div className="mb-4 text-center text-sm font-medium text-green-600">
-                    A new verification link has been sent to the email address you provided during registration.
-                </div>
-            )}
+      {status === 'verification-link-sent' && (
+        <div className="mb-4 text-center text-sm font-medium text-green-600">
+          A new verification link has been sent to the email address you provided during registration.
+        </div>
+      )}
 
-            <form onSubmit={submit} className="space-y-6 text-center">
-                <Button disabled={processing} variant="secondary">
-                    {processing && <LoaderCircle className="h-4 w-4 animate-spin" />}
-                    Resend verification email
-                </Button>
+      <form onSubmit={submit} className="space-y-6 text-center">
+        <Button disabled={processing} variant="secondary">
+          {processing && <LoaderCircle className="h-4 w-4 animate-spin" />}
+          Resend verification email
+        </Button>
 
-                <TextLink href={route('logout')} method="post" className="mx-auto block text-sm">
-                    Log out
-                </TextLink>
-            </form>
-        </AuthLayout>
-    );
+        <TextLink href={route('logout')} method="post" className="mx-auto block text-sm">
+          Log out
+        </TextLink>
+      </form>
+    </AuthLayout>
+  );
 }
