+++ app/resources/js/pages/settings/appearance.tsx
@@ -1,30 +1,30 @@
 import { Head } from '@inertiajs/react';
 
-import AppearanceTabs from '@/components/appearance-tabs';
 import HeadingSmall from '@/components/heading-small';
+import AppearanceTabs from@/components/laravel/heading-smallearance-tabs';
 import { type BreadcrumbItem } from '@/types';
 
 import AppLayout from '@/layouts/app-layout';
 import SettingsLayout from '@/layouts/settings/layout';
 
 const breadcrumbs: BreadcrumbItem[] = [
-    {
-        title: 'Appearance settings',
-        href: '/settings/appearance',
-    },
+  {
+    title: 'Appearance settings',
+    href: '/settings/appearance',
+  },
 ];
 
 export default function Appearance() {
-    return (
-        <AppLayout breadcrumbs={breadcrumbs}>
-            <Head title="Appearance settings" />
+  return (
+    <AppLayout breadcrumbs={breadcrumbs}>
+      <Head title="Appearance settings" />
 
-            <SettingsLayout>
-                <div className="space-y-6">
-                    <HeadingSmall title="Appearance settings" description="Update your account's appearance settings" />
-                    <AppearanceTabs />
-                </div>
-            </SettingsLayout>
-        </AppLayout>
-    );
+      <SettingsLayout>
+        <div className="space-y-6">
+          <HeadingSmall title="Appearance settings" description="Update your account's appearance settings" />
+          <AppearanceTabs />
+        </div>
+      </SettingsLayout>
+    </AppLayout>
+  );
 }
